ai_detector/requirements.txt
numpy
pandas
Pillow
matplotlib
scikit-learn
torch
torchvision
PyYAML
joblib
